refactor(web): relocate generic hooks from pages into the shared hooks directory
