+++ web/src/pages/builtin/devices/DevicesPage.tsx
@@ -2,13 +2,12 @@ import React, { useState, useCallback, useMemo, useEffect } from 'react';
 import { useSearchParams } from 'react-router-dom';
 import { Button, Icon } from '@gravity-ui/uikit';
 import { Plus } from '@gravity-ui/icons';
-import { useSearchParamHelpers, useListNavigation, usePageContribution } from '../../../hooks';
+import { useSearchParamHelpers, useListNavigation, usePageContribution, useUnsavedChangesBlocker } from '../../../hooks';
 import { PageLayout, PageLoader, EmptyState, CommandPaletteHeader } from '../../../components';
 import type { DeviceType } from '../../../api/devices';
 import type { LocalDevice } from './types';
 import { useDeviceCounters } from '../../../hooks';
 import { useCounterHistory } from '../../../hooks/useCounterHistory';
-import { useUnsavedChangesBlocker } from '../_shared/lane-editor';
 import type { Command, RowAdapter, ShortcutSection, PagePaletteContribution } from '../../../components/command-palette';
 import {
     DevicesList,
