+++ web/src/pages/modules/acl/AclPage.tsx
@@ -3,9 +3,8 @@ import { Button, Icon, Label } from '@gravity-ui/uikit';
 import { Funnel, Pause, Play, Plus } from '@gravity-ui/icons';
 import { useNavigate, useSearchParams } from 'react-router-dom';
 import { PageLayout, PageLoader, ConfigTabStrip, BulkBar, SearchInput, EmptyPagePlaceholder, RowCountDisplay } from '../../../components';
-import { useSearchParamHelpers, usePageKeyboardShortcuts, useDirtyConfigSet, useConfigQuerySync, usePageContribution } from '../../../hooks';
+import { useSearchParamHelpers, usePageKeyboardShortcuts, useDirtyConfigSet, useConfigQuerySync, usePageContribution, useTabCycle, useUnsavedChangesBlocker } from '../../../hooks';
 import { useAclDraft } from './useAclDraft';
-import { useUnsavedChangesBlocker } from '../../builtin/_shared/lane-editor';
 import type { Rule } from '../../../api/acl';
 import { ActionKind } from '../../../api/acl';
 import type { RuleItem, RuleDraft } from './types';
@@ -20,7 +19,6 @@ import { AddConfigModal, DeleteConfigModal, BulkDeleteModal, CommandPaletteHeade
 import { DRAWER_TRANSITION_MS } from '../../../components/draft';
 import type { Command, RowAdapter, PagePaletteContribution } from '../../../components/command-palette';
 import { buildConfigCommands } from '../../../components/command-palette';
-import { useTabCycle } from '../../_shared/useTabCycle';
 import '../../../styles/chrome.scss';
 import './acl.scss';
