+++ web/src/pages/modules/fwstate/FWStatePage.tsx
@@ -14,18 +14,15 @@ import {
 } from '@gravity-ui/uikit';
 import { CircleInfo, Plus } from '@gravity-ui/icons';
 import { useNavigate, useSearchParams } from 'react-router-dom';
-import { useSearchParamHelpers, usePageContribution } from '../../../hooks';
+import { useSearchParamHelpers, usePageContribution, useContainerHeight, useTabCycle, useUnsavedChangesBlocker } from '../../../hooks';
 import { API } from '../../../api';
 import { Direction, type FwStateEntry, type ListEntriesRequest, type MapStats } from '../../../api/fwstate';
 import { ConfirmDialog, ConfigTabStrip, PageLayout, PageLoader, EmptyPagePlaceholder } from '../../../components';
-import { useContainerHeight } from '../../../hooks';
-import { useUnsavedChangesBlocker } from '../../builtin/_shared/lane-editor';
 import { ipAddressToString, isValidIPAddress, parseIPToBytes, stringToIPAddress, type IPAddressWire } from '../../../utils/netip';
 import { parseMACToBytes } from '../../../utils/mac';
 import { formatBytes, toaster, compareNatural } from '../../../utils';
 import { AddConfigModal, DeleteConfigModal, CommandPaletteHeader } from '../../../components';
 import { SaveIcon, TrashIcon } from '../../../components/draft';
-import { useTabCycle } from '../../_shared/useTabCycle';
 import type { Command, PagePaletteContribution } from '../../../components/command-palette';
 import '../../../styles/chrome.scss';
 import './fwstate.scss';
